pybricks/nxtdevices/ColorSensor: support color_map
This makes the NXT Color Sensor work like the Powered Up
Color Sensor Sensor API. Since it just adds the color_map()
method, this does not break backwards compatibilty.

While that is useful in itself, the main reason for doing this is
so we can debug all color features on ev3dev, which saves a lot
of time since firmware upgrades are not required.