sys/battery: fix Control+ hub early shutdown
Since 33191c73, Control+ would shutdown immediately after booting
because it thought that the battery voltage was low. This is fixed
by making the initial average voltage no less that the low voltage
shutdown voltage. If the actual voltage really is lower, the average
will quickly become lower and trigger the shutdown. But if the voltage
really is higher, it won't take so long for the average to increase to
the actual voltage.