pbio/serial: move to package
Legacy platform-specific drivers such as those in drv/ev3dev_stretch/serial are being to be removed. This prepares for that by removing its platform specific pbio component also.

This commit just moves it over to the package without code changes. Next up, it can
be refactored significantly because it is much simpler to wait for data in a (nearly) non-blocking way at the module level.