pbdrv/ev3dev_stretch/serial: move to pb_util
Legacy platform-specific drivers such as those in drv/ev3dev_stretch/*
are all being (re)moved. This is one of the last remaining drivers.

Quite like pb_device, it is moved to pb_util, where we can have both
an ev3dev and stm32 implementation in the future. This moves the code as-is, with pbdrv_ renamed to pb_