pybricks/iodevices/UARTDevice: simplify read logic
This drops the legacy pbio pattern of calling a function until it no longer
returns PBIO_ERROR_AGAIN. Since this part of a MicroPython module,
it can simply wait using mp_hal_delay_ms. This way, we don't need any
state variables to keep track of which start/busy/stop state we are in.