pbio/sys/prime_hub: reset lightgrid on start/stop
This is a very minimal UI to indicate whether a program is runs or not.

The run status is meant to be indicated by a fading bar that always
moves right, like the familiar indefinite progress bar, but patterns
are not yet implemented. Instead just display a static bar for now.