src/light: new color light implementation
This replaces pbio_light_* with pbio_color_light_*. The new
implementation is more generic and allows for instance-specific
callbacks so that it can be used with an arbitrary light.

Helper functions for light on/off are moved from the low-level LED
driver to the new color light interface.

The sys/light implementation is expanded to replace the old pbio_light_*
implementation for the system status light. Arbitrary lights don't need
to know the different between user programming running and not, so this
logic is moved to sys/light.

A more generic light animation implementation is given. However it is
not used anywhere yet, so this breaks the breathe pattern that was used
when a user program is running. For now, it is replaced with solid
green. The animation backend can also be shared with other light
interfaces like the light array/grid.

The flashing white light on power down is also removed for now since it
was in the low-level reset driver and didn't work on some platforms
anyway.

The ev3dev-stretch status light implementation is reworked with the
intention of eventually giving full control of the LEDs to brickd
rather than poking the sysfs attributes directly. Brickd doesn't have
any status light commands yet, so for now we just let it have control
until the first time user program changes the status light.