sys/battery: fix Control+ showing low battery on boot
The Control+ hub reads low battery voltage via the ADC when it boots for
some reason. We have already worked around it so that the hub doesn't
shut down immediately at boot time but it is still showing a low voltage
warning at boot, so we adjust the initial voltage value even higher to
avoid the low voltage warning.