bricks/*/main.py: print version and wait
This updates all main.py scripts.

uos.uname() is no longer used to get the version, so we print pybricks.version.

It also adds a long wait and a "Hello, World" message, which lets us quickly test the UI without uploading a new program. Once this is finalized, we can remove it and settle on a different main.py program, or none at all.