pbio/drv/led/led_pwm: fix gamma correction
Gamma correction was not being applied which caused higher brightness
values to be visually indistinguishable.

For example, `hub.light.animate([(30, 100, b) for b in range(100)], 50)`
would cause the light to seem to increase in brightness for short time
then stay at max brightness for a long time instead of continuously
changing brightness.