light/animation: add stop all function
This reworks the animation module so that we can stop any background
animations that are still running when a user program stops.

This is needed since MicroPython can stop in the middle of an unhandled
exception, so there is no way to guarantee that individual stop methods
will be called. Therefore reference counting can't work.

Reference counting is replaced by a linked list to keep track of all
running animations so that they can be stopped individually or all at
once.