src/color: introduce compressed HSV type
This adds a new type to hold compressed HSV data. With the 360/100/100
value ranges, it is possible to fit an HSV value into 24 bits without
loss.

Replacing the existing HSV value type with the compressed on would
greatly increase the firmware size since additional instructions would
be needed any time one of the fields is accessed. So we only use the
compression when there would be a significant savings in space like
when arrays of HSV values are used.