pybricks/util_pb/color_map: drop existing mapping
Previously, the color_map was a user-specified mapping from Color enums to hue / values. Since the introduction of the new Color type, the Color already *is* an HSV color.

In this case, the user just has to provide a list/tuple of detectable/desired colors. Then we can discretize the HSV measurement to the nearest match in the set of colors.

This commit prepares the framework for this, and leaves open the actual mapping from measurements to discrete colors.