pybricks/util_pb/color_map: reimplement mapping
The matching color is the one that differs the least from the measurement.

A simplistic cost function is currently implemented so that the sensors keep working. At the moment, it is similar to the basic mapping that we used to have, which could only detect either fully saturated colors or non-colors.