pybricks.pupdevices.PFMotor: fix red motor on channel 1
After every message, we send a value to the sensor that does not do anything. This is a necessary workaround so we can force resending of data to improve reliability.

This fix changes that bad value from 0 to INT16_MAX, because 0 had the side effect of stopping the red motor on channel 1.