sys/light: rotate indication patterns
Since introducing the BLE advertising pattern the status light could be
off when the hub first boots since the first step of the pattern is to
turn the light off for 1 second. This makes it look like the hub did
not turn on. By rotating the patterns so that the first step always
turns the light on, we can avoid this problem.

Fixes pybricks/pybricks-micropython#16