tools/build_llsp: install Pybricks as user project
This is a proof of concept to install Pybricks firmware using the
official SPIKE Prime app. This script builds an llsp file that can be
opened and run within the app. Since the file is about 200 kb,
downloading takes about 30 seconds, during which no progress bar is
currently shown.

For now, it replaces the existing firmware so restoring still has to be
done via DFU. Ultimately, this technique can be used to install Pybricks
alongside the official firmware.