pybricks/util_pb: modify rgb to hsv conversion
This adds in a few modifications to make HSV match reality somewhat better. In the long run, this needs to be done properly similar to how HSV to RGB is calibrated. That way, we would also distinguish orange from red better.

At the moment, all sensors have the exact same mismatch, so we use a shared correction function. Better results may be obtained by tweaking it per sensor.