pybricks.experimental: add getchar()
We frequently get requests for some form of augmented remote control: The robot is mostly autonomous (taking care of time critical stuff), while the user sends high level commands like which direction to drive in.

While a complete form of hub-to-hub communication is still a long way off, we can already enable the user to access standard-in in a non-blocking way. This way, users
can create robots that respond to characters typed in the Pybricks Code i/o pane.

This is much more flexible than input(), which is blocking, requires and end of line character, and echos.