pybricks.media: add ImageMatrix
Similar to SoundFile and ImageFile for ev3dev, this provides
an ImageMatrix enum for commonly used icons on the light matrix display.

Images are generated on the spot from a single integer to save space.
They are returned as a Matrix, so it is easy to change the brightness
or make composites.