platform/technic_hub: move red, green LEDs to TIM2
This moves the red and gree LED PWMs to TIM2. This is done so that we
can change the motor PWM period independently of the LED PWM period.
This means we can't use the pin mux to directly connect the TIM2 outputs
to the proper pins, so we use an interrupt to poke the GPIOs instead.