platform: increase motor PWM frequency 10x
This increases the motor PWM frequency on all platforms from the 1/1.2
kHz range to the 10/12 kHz range. This reduces the motor noise when the
PWM is active. This is similar to EV3, which uses a 13.2 kHz PWM for
motors. This comes at a small cost of a slightly less linear duty cycle
vs. torque curve.