drv/core: implement blocking driver init and deinit
This adds a busy loop to pbdrv_init() to wait until drivers with async
init are ready. Currently only pwm drivers do this. However most drivers
should do this eventually to allow for dependencies to properly resolve.

This also adds a new prdrv_deinit() function. As mentioned in the
comments in the code, we really shouldn't need this, but for now it
serves as a quick fix to get lights to turn off during power off before
the power button is released.