pbio/motor_process: refactor
This cleans up and renames the previous motorpoll logic.

Instead of using PBIO_ERROR_AGAIN in different places to tell whether a motor should be polled, we now use a dedicated connected flag on each servo.

Drivebases should eventually have something like this as well. For now, this is taken care of by the update function itself, which skips control if the drivebase is not in use.

There is a slight braking change for EV3, which is that the motor reset now always puts the port back in tacho mode on startup, because the motor reset calls the servo setup function to check if a servo is connected.