bricks/stm32: add ReadMe_OSS.txt per hub
This adds a ReadMe_OSS.txt file for each STM32-based hub. This file
includes the licenses and copyright statements for binary distribution.

These files were created using Fossology after manually auditing the
source code with the assumption that the binary is built using the
default configuration for each hub.