bricks/movehub: reduce heap size
We were starting to run out of RAM for stack and statically allocated variables on Move Hub, which has only 16K of RAM.

This works around it by reducing the heap for the MicroPython user program.

This fixes #29, but only temporarily. We may need to disable some features on Move Hub instead or optimize some of the drivers, such as the motor controllers.