pbio/servo: reinstate logger objects
All logs were moved to the control onbjects previously, but it is more helpful to log servo states separately.

For example, a drivebase has data to log for its controller, but it has two separate servos, each with a model observer.

This also makes it more natural to log passive states when no control is active.