drv/bluetooth: new BTStack/CC256x/STM32 driver
This adds a new Bluetooth driver for SPIKE Prime/MINDSTORMS Inventor
that uses BluetKitchen BTStack for the Bluetooth stack.

The driver consists of:
- A btstack_control_t to control the nSHUTD pin on the CC2564C.
- A btstack_run_loop_t to integrate BTStack into the Contiki run loop.
- A btstack_uart_block_t to integrate the STM32 UART with BTStack.

The CC2564C init script was automatically generated by BTStack by
running:

    make -C lib/pbio/drv/bluetooth -f ../../../btstack/chipset/cc256x/Makefile.inc BTSTACK_ROOT=../../../btstack bluetooth_init_cc2564C_1.4.c

Bluetooth advertising data is configured similar to the other hubs with
the Pybricks service UUID in the advertising data and "Pybricks Hub"
for the local name in the scan response data.

The nRF UART service is implemented using the BTStack implementation
and replaces the USB serial on the Prime/Inventor hub.

ENABLE_CC256X_BAUDRATE_CHANGE_FLOWCONTROL_BUG_WORKAROUND doesn't seem
to be working, so we can't currently change the baud rate, so the
connection is a bit slower than it could be.