bluetooth/bluetooth_btstack: don't wait in process loop
This removes the blocking wait in the main process loop. This could
cause events to be missed.

Also fixed the TODO by replacing PROCESS_EVENT_MSG with process_poll()
and removed gap_advertisements_enable(false) since that happens
automatically during connection and added a wait to be sure that the
chip is really off before resetting.