pbio/servo: only reset used servos
After a program completed we were resetting all hbridges for simplicity, but this was also turning off sensors that need power, like the Powered Up Color Sensor.

This ensures that we only reset those motors for which we have a dc motor reference, which could only be initialized if there was a tacho.

We will still need a mechanism for resetting dc motors.