pybricks.common.Keypad: convert module to class
This converts the legacy button module to a keypad class. This follows the API specification to make it consistent with the other hub attributes.

By making it a class, it can have attributes, states, and configurable settings such as the number of presses until a KeyboardInterrupt is raised.

The pressed method currently works the same as before, but now we can more conveniently address differences between hubs by passing the available buttons when initializing the class object.