pybricks.common.Keypad: cache button combinations
Scripts such as

while not any(hub.buttons.pressed()):
    pass

were consuming heap quite quickly because it kept allocating new lists of buttons.

This changes ensure that unique combinations are cached as they are pressed, so we allocate at most 2^num_buttons of button combinations.

This also changes the return type from list to tuple, since the key combinations may not be modified.