treewide: use void when no parameters
Until now, we have been a bit lax in omitting the void keyword when
functions have no parameters. However, in C, an empty parameter list
means that the parameters have not been defined yet whereas the void
keyword means that there are no parameters.

This adds void everywhere to be more explicit and for consistency.