pybricks: fix non-extern variables in header files
In C, function definitions are implicitly extern, so the extern keyword
is not needed. But variables are not implicitly extern, so the extern
keyword is always needed.

Prior to GCC 10, the `-fcommon` flag was set by default which silently
ignored this mistake.

Fixed by adding extern everywhere. This also uncovered a bug where
`pybricks.parameters.Icon` was included on all hubs but would have
resulted in a crash when used on hubs where it should have been omitted.

Issue: https://github.com/pybricks/support/discussions/194