platform/technic_hub: fix race condition in TIM2 irq handler
This fixes a race condition where an interrupt status flag is set while
the interrupt handler is running. We were clearing all flags each time
the handler ran but now we only clear the flags we have handled.

This was triggered when CCR1 was set to 3, causing the green LED to turn
on and stay on all of the time.

Fixes: pybricks/support#224