platform/prime_hub: fix dual boot with GCC10
GCC10 started doing tail call optimization on jump_to_other_firmware().

GCC9:

        080e8c08 5b 68           ldr        r3,[r3,#offset PTR_Reset_Handler+1_080c0004]
        080e8c0a 98 47           blx        r3=>Reset_Handler
        080e8c0c 16 b0           add        sp,#0x58
        080e8c0e 70 bd           pop        { r4, r5, r6, pc }

GCC10:

        080e8d08 5b 68           ldr        r3,[r3,#offset PTR_Reset_Handler+1_080c0004]
        080e8d0a 16 b0           add        sp,#0x58
        080e8d0c bd e8 70 40     pop.w      { r4, r5, r6, lr }
        080e8d10 18 47           bx         r3=>Reset_Handler

This causes a lockup since we are resetting the stack pointer before
this which causes the stack pointer to get set to a value outside of
the addressable memory.

Fix the issue by using inline assembly to do the jump instead of relying
on GCC to do the right thing with a function call.