drv/bluetooth/cc2640: fix Read by Group Type Response
When services are enumerated, the connecting device usually starts with
a Read by Group Type Request for handles 0x0001 to 0xffff. However,
the first handle, gatt_service_handle, seems to be 0x0009, so this
would result in an application error response since no if case matched
0x0001.

This fixes the problem by using inequalities instead of strict equals.

Also, when all groups have been enumerated, it is supposed to send an
attribute not found error, not an application error, so this is fixed
too.