bricks/stm32: enable uio, usys and uselect modules
This enables access to stdio as a stream object. This will allow more
advanced use cases. Also, stream objects will be the base for generic
I/O port UART communications in the future.

This can be used as a replacement for the experimental getchar() function:

    from usys import stdin
    from uselect import poll

    p = poll()
    p.register(stdin)

    def getchar():
        return ord(stdin.read(1)) if p.poll(0) else None

It can also be used to create an event loop:

    from usys import stdin
    from uselect import poll

    hub = TechnicHub()

    p = poll()
    p.register(stdin)

    while True:
        if p.poll(10):
            print(ord(stdin.read(1)))

There is not enough flash memory left to enable these on Move hub.