pbsys/sys: drop stdio functions
This removes the pbsys stdio abstraction layer. Instead, we connect the
MicroPython stdio HAL directly to the pbsys Bluetooth.

The use of the light weight ring buffer library, despite its name, adds
a bit of size to the firmware. However, we can make more optimizations
to that library and use it in more places to reclaim some of that space.