sys/bluetooth: fix multiple disable/enable of Pybricks service notifications
This allows the Pybricks service notifications to be enabled and disabled
multiple times during a single connection. This is "good" behavior anyway
but it is required for us to work around https://crbug.com/1170085.

This also required some modifications to the BTStack driver to get
proper callbacks when notifications are disabled.