pbsys/user_program: move stop button logic
This moves the stop button logic into the user program module. It makes
sense to put it here since it only has an effect when a user program is
running.

We also add the capability for changing the stop button and allowing a
combination of buttons for the stop button. So we have to change it to
be polled instead of event driven.

Issue: pybricks/support#272