drv/bluetooth/btstack: update to develop branch
This pulls in some changes so that we can detect when notifications
of the Nordic UART service are disabled. This is needed to work around
a Chromium Linux bug where we have to disable and enable notifications
more than once while the device is connected.

There were also some changes in logging that had to be fixed.