sys/bluetooth: add status update timer
This adds a timer to send status updates periodically in addition to
when the status changes. Since there is no guarantee that notifications
are received, we just keep sending them to be sure.

In the future, we could consider sending an ack back to the hub indicate
that the status change was received so that we don't have to do this.

Issue: https://github.com/pybricks/support/issues/287