pybricks.hubs: consolidate reset functions
We had individual reset functions for each reset type. However, these
are not frequently used, so we can consolidate them to a single function
to save on flash memory. This will also allow for addition reset actions
to be added in the future without requiring addition MicroPython
functions to be added.

Also refactor duplicate code while touching this.