drv/reset/stm32: move getting reason to init
The RCC_CSR flags persist across reboots, so we could end up in a
situation where both the software and watchdog flags were set and
would not be able to tell which what the true reason.

By moving this to the init, we can get the reason and reset the flags
so that the flags will always be cleared for the next reboot.