bricks/primehub: Revisit dual boot approach.
The experimental web installer has been giving promising results. This means we can drop support for the somewhat cumbersome installation via the official apps.

We will proceed with a mix of both ideas, taking the best of both. That means we'll upload files via the web, but run the installation locally similar to how we ran it via the app before. This is the fastest and most reliable solution.

We'll include the installation files in the ZIP firmware build so that in the future we can match this to specific LEGO firmware versions that we have tested succesfully.