pbio/uartdev: Move external type_id setter.
The original comment suggested that this is used to let external devices know that a device is ready, but this isn't strictly true. It may still be reporting garbage data at this point. So, we move it to the point where it has received one good data message.

It would be better to have a proper function to check this, but this can be done as part of the bigger lpf2 driver overhaul in a future release.

https://github.com/pybricks/support/issues/361