drv/bluetooth: add PnP ID to Device Info Service
This adds a PnP ID characteristic to the Device Information Service.
This will allow us to identify the type of hub. The IDs come from LWP3.
The Technic Large hub will read the device variant from the bootloader
flash memory area and use that as the product version so that we can
tell the difference between SPIKE Prime and MINDSTORMS Robot Inventor.