drv/bluetooth: add special location for hub name
This modifies the advertising data and device name characteristics to
use the hub name from a special place in the firmware flash memory.
The offset and size of this location is stored in the firmware metadata
json file to allow modifying the hub name before flashing the firmware.
The name must be zero-terminated since we are using strlen(), so the
max size in the metadata includes the zero-termination byte.

Issue: https://github.com/pybricks/support/issues/52