tools/build-dual-boot-bin: fix placement of checksum
Since `firmware-dual-boot-base.bin` doesn't reserve space for the
checksum, the 'build-dual-boot-bin` tool was writing over the last 4
bytes in the firmware. This became obvious after 53d49742f since it
contained exactly 4 0s after "Pybricks Hub" at the end of the firmware.
This caused `strlen()` to try to read past the end of the flash memory
(uninitialized flash memory is set to 0xFF so strlen() never found the
zero-termination of the string).

Fixes: #56