sys/bluetooth: keep Bluetooth chip on when not connected
This changes the logic in sys/bluetooth to keep the Bluetooth chip on
while a user program is running, even when not connected to a central.
This allows the program to initiate a connection to a Bluetooth device
like the Powered Up remote control.

This also required implementing a new drv/bluetooth function to stop
advertising when a user program starts.