drv/reset: new pbdrv_reset_power_off() function
This adds a new pbdrv_reset_power_off() function that is intended for
flipping the switch on the power supply. This will allow us to turn
the power off but keep software running. This is needed since hubs may
not actually turn off due to the power button being pressed or the USB
being plugged in.