sys/main: Improve handling of hub shutdown.
Since hubs may not actually power off on shutdown, (e.g. if the button
is being pressed or USB is plugged in), we have to have special handling
to make it appear as if the hub has powered off so that the user will
know to release the button.

A new mechanism to forcibly abort user programs is needed since
"stopping" a user program may not always happen (e.g. it is possible
to catch SystemExit in MicroPython).

- Add new sys/main module with mechanism for aborting a user program
- Add a new hub status flag for requesting shutdown
- Update sys/bluetooth to power off the Bluetooth chip when shutting down
- Update sys/light to turn of the hub status light when shutting down
- Update sys/light_matrix to turn off all of the lights when shutting down
- Update sys/hmi to allow turning hubs with battery charger back on
- Add event loop hook for sys/main to hook into to abort user program

Issue: https://github.com/pybricks/pybricks-micropython/issues/36