drv/ioport_lpf2: add move hub shutdown quirk
This adds a configuration option to handle a quirk of the BOOST Move
hub. For some reason, this hub turns back on when the power button is
released after shutdown if the port VCC is turned off.

Also added PROCESS_EXIT() to the exit handler just to be safe while we
are touching this (otherwise device connection manager might run one
more iteration).

Fixes: https://github.com/pybricks/support/issues/386
Fixes: 1473c4ba649fd38d07731c6104f66bf43b083d38