BlueNRG: drop return value on *_begin() functions
These functions always return success and we don't check the return
value, so we can removed the return value to reduce the firmware size.

On the few functions that did return an error, it was just arg checks,
so these are changed to assert() since we don't anticipate invalid args
at runtime.

Issue: https://github.com/pybricks/pybricks-micropython/issues/59