src/task: add cancel function
This adds a function to cancel tasks. This sets the cancel flag and
runs one iteration. If we didn't run an iteration, then it could take
some time for an event to trigger running the task before the cancel
was actually handled (or it might never be handled if there were no
events!).

This change also means that a task can complete without being removed
from a queue, so we have to add a check for this to ensure that the
protothread isn't accidentally restarted by running an iteration after
the task has completed.