pybricks.pupdevices.ColorDistanceSensor: use COL_O mode
This changes the light.on() implementation to use the COL_O mode. Since
this function changes the mode anyway, we might as well use the mode
specifically designed for it. Also since the Color type is no longer
an enum, we just use the closest color in the HSV wheel in case an
inexact value is used.