drv/bluetooth_stm32_cc2640: use write without response
This modifies the pbdrv_bluetooth_write_remote() function in the CC2640
driver to use write with response instead of write without response.
This ensures that the remote device received the message before the
task ends.

This fixes issues like messages at the end of a MicroPython program not
being sent before the remote device is disconnected when the program ends.

After this commit, connecting to a Handset fails with RuntimeError.
This is expected due to a separate bug where remote_lwp3_char_handle
is being overwritten (and actually shows that one of the error paths
is working correctly).

Issue: https://github.com/pybricks/support/issues/432