drv/bluetooth_stm32_bluenrg: fix hci_recv_res()
This fixes the implementation of the hci_recv_resp() callback for the
BlueNRG-MS library. The *_end() functions expect that commands that
end with a EVT_CMD_COMPLETE event to skip over the common header
(evt_cmd_complete struct) when unpacking the response. However, other
commands that end with other events expect the full response data.

Up to this point, we were probably ignoring the response from all
commands that end with EVT_CMD_STATUS so we didn't notice the problem.