pybricks.pupdevices.Remote: set light to Blue on connect
This set the light to blue when connecting to the handset. Changing the
mode to RGB will make the LED show the last used color, so we need to
set a color to get consistent behavior.

Several functions were moved to avoid forward declarations.

Issue: https://github.com/pybricks/support/issues/425