drv/bluetooth: use address type when connecting
The LEGO Mario hub uses a random BLE address, so we need to store this
information from the advertising data and use it when connecting.

We can also drop the Mario attribute handle hack since Mario seems to
be working without it now (this may have been fixed earlier in a4a85dd3
or it could be due to this change but wasn't tested until now).