drv/bluetooth_stm32_cc2640: handle macOS 12 request
It looks like macOS 12 issues a new request we have't seen before:
a Read By Type Request for the Device Name UUID. This was causing
a connection failure in Pybricks Code. Since the hub never sent a
response to this request, the OS would timeout and disconnect.

This fixes the issue by handling the request. Plus an error response
is now sent for unhandled UUIDs (as per the Bluetooth spec) to prevent
similar problems in the future.

Fixes: https://github.com/pybricks/support/issues/489