pybricks.common.Control: Restore arg order.
The integral_range argument was dropped in
3f0042f988127a2d88031f920b7f9c943c301f61.

While this setting was virtually never used, the pid()
method is used in some example projects. Typically, all 5 parameters
are unpacked to easily read the existing PID parameters. Removing one
argument would break these projects.

This commit restores the order and number of arguments to avoid a
breaking change. The setting may be reserved for future use instead.