drv/clock: move clock driver out of contiki
This refactors the code base to use pbdrv_clock_* instead of clock_*
from contiki. This is a step towards eliminating contiki from the
public API, which will be nice if we ever have PBIO as a stand-alone
C library.

It also simplifies the code a bit since it gets rid of clock_from_msec()
and clock_to_msec() which were NOPs in the PBIO library anyway.

clock_delay_usec() is moved into the MicroPython code (as
mp_hal_delay_us()) since we want to avoid any blocking functions
in PBIO.