bricks/stm32/main: Save programs on Primehub.
This hub has dedicated storage space for user programs. Now the
downloaded program is stored so it can be started with the button
later. It is saved to /_pybricks/main.mpy.

This needs to be reworked extensively to make file access clean and
non-blocking, but it makes the hub much more useable for now.