pbio/drivebase: Use measured position for feedback.
Using the estimated position makes motion smoother, but it
makes it harder to reach the target angle at the end.

However, since overal behavior is a bit smoother due to previous
improvements to the observer model gain, we can set this
back to position feedback to get the right balance of both.