pbio/control: Control by voltage.
This replaces duty cycle by voltages as the lowest-level control signal as far as pbio is concerned.

This pushes implementation-specific details of duty cycle further down. This is currently pushed down to pbio_dcmotor, but we could push it even further down to the pbdrv level.

This removes implicit asumptions that duty cycle has 10,000 increments which was ambiguously used in some of the observer calculations.

This also ensures that we don't need the battery voltage for observer calculations.