pbio/servo: Drop check for state getter.
This check existed to avoid unnecessary state lookups in tight loops, which causes unnecessary I/O operations in some implementations like ev3dev.

But for the platforms we are currently targeting, this just involves reading a cached value, so there is little harm in doing so.

For other platforms, we could also consider having the driver return a cached value also if a value isn't older than the platform-specific update rate.