pbio/servo: Fix actuation type from control.
Now we no longer ignore the actuation type returned by the controller.

This fixes a longstanding bug where the _then_ argument (e.g. Stop.COAST) was essentially always ignored.

This was introduced when adding the experimental observer early 2021 without updating underlying abstraction layers.

It was not that noticable because on Powered Up motors, coast is fairly close to a near-zero duty cycle (brake state).
Moreover, Stop.HOLD is the default, which was not affected by this bug.