pbio/drivebase: Simplify acceleration setter.
It was being doubled in pbio, then halved again in the MicroPython
module for a subset of cases.

We drop the duplication and halving. Instead, we double the drive base
acceleration in the drive() method to keep unchanged behavior for now.
It would be better to have a dedicated setter for line following.

Additionally, turning was halved once more for turn rate and turn
acceleration during turn(). This is dropped entirely. With recent
control improvements, this special case is no longer needed/justified.