drv/charger: initial support for battery charger
The SPIKE Prime and SPIKE Essential hubs have a MPS MP2639A battery
charger chip for charging the battery from USB. This adds a new charger
driver framework and implements monitoring the charge status and the
charge current (with some TODOs).

There are also some quirks required for SPIKE Prime since not all of the
charger inputs and outputs are connected directly to the MCU but rather
go through other external peripherals.