drv/resistor_ladder: new driver
This adds a new driver for resistor ladders as found in SPIKE Prime.
These use an analog input to multiplex multiple digital inputs.

Previously, this was implemented as part of the button driver, however
not all of the input are buttons. So, we create a new driver to make
the input mechanism more generic so it can be shared with the battery
charger driver.