drv/charger_mp2639a: detect fault
This implements fault detection for the MP2639A by monitoring the CHG
signal status. If there is a fault, this signal will oscillate at 1Hz,
50% duty cycle. So we have to constantly sample this signal to see if
this is the case.

This also allows us to remove the error return from
pbdrv_charger_get_status() since it now just reads a variable
and I/O is done in the background.