pbio/motor_process: Refactor error propagation.
The motor process has currently no way of propagating errors. If errors occur in the control loop or a motor is unplugged, this can lead to undefined behavior.

This is a work in progress aimed at addressing this.

The motor process now sets an error flag if something goes wrong. This flag is polled by pbsys, in the same way that the button is monitored.

We also internalize the servo->connected flag instead of using external getters and setters for it.