bricks/stm32/configport: Enable ujson.
This is quite useful for hubs that can read and write from
external storage. It is currently enabled on all hubs with
the PYBRICKS_STM32_OPT_EXTRA_MOD flag.

It is also used for the firmware restore process, so we can
parse JSON files to check meta data without reinventing
the wheel.