pbio/control: Introduce system state type.
Most of the control functions pass a large number of arguments
in and out, which quickly gets hard to follow. These arguments
will now be grouped by several types.

In this commit, we introduce the pbio_control_state_t, which contains
the physical and estimated state of a generic point-mass like system,
such as a simplied servo motor, a drivebase mass, and a drivebase
inertia.