pbio/trajectory: Change speed computation as dependent variable.
This changes how the dependent variable speed is re-computed for
a trajectory with a known final time and angle.

Before, the new trajectory had the same acceleration values but a
new speed, and new time values for t1 and t2.

The problem there is that t1 and t2 of a follower trajectory (for which
we will use this new trajectory) are generally different from t1 and t2
of the leader. This makes curved motion look less smooth because it
appears as if linear and rotational acceleration are different.

In this new version, we allow a0 and a2 to be free, and fix t1 and
t2 to be the same as the leader trajectory. We still solve for the
target speed as well.

Intuitively, this function now reduces the speed by stretching out
the time line for the whole maneuver, adjusting speed and acceleration
such that the integral of the speed curve still equals the same angle.

Since the function is only used to stretch (not compress) a trajectory
that already satisfies maximum speed constraints, no further constraints
are needed to stretch it, since the new move is always slower.

As a bonus, this also avoids the use of sqrt(), which avoids a problem
where it's argument would overflow for large angles or radii.