pbio/control: Use functions to read control type.
We will shortly be adding a new control type that is very similar to angle control. If we use these functions to check types, we can just modify those functions as we add new types, without having to do this everywhere.

It also makes the code a bit easier to read.