pbio/drivebase: Ensure there is a leader and follower.
If the two trajectories already had equal length, we were skipping the
leader/follower determination, but this is still needed to obtain the
desired synchronization result.

In principle we could still skip the stretching step (which is why the
deleted line existed in the first place), but this is an inexpensive
step that is quite unlikely to be executed anyway.