drv/ioport_lpf2: fix improper device detection
Occasionally, I/O devices would be improperly detected when plugged in
while the hub is already turned on. For example, a medium motor (id 1)
would be detected as a light (id 8). This points to a pull down resistor
on the ID2 pin not being properly detected. This could be due to some
internal capacitance causing it to take longer than 2 ms for the voltage
to drop low enough to be detected as logic low after voltage is changed
from high to low on the pin. By driving the detection pin low, we give
it a boost in the right direction to improve the chances of proper
detection.

Fixes: https://github.com/pybricks/support/issues/500